refactor(parsers): restructure parser module
